--- app/src/services/system-checks.ts
@@ -1,237 +1,0 @@
-import { ServerStatusResponseData } from '@/bcsc-theme/api/hooks/useConfigApi'
-import { BCSCBanner } from '@/bcsc-theme/components/AppBanner'
-import { IdToken } from '@/bcsc-theme/utils/id-token'
-import { BCDispatchAction } from '@/store'
-import { BifoldLogger, ReducerAction } from '@bifold/core'
-import { TFunction } from 'i18next'
-import { Dispatch } from 'react'
-
-export type SystemCheckStrategy = {
-  /**
-   * Runs the startup check.
-   *
-   * @returns {Promise<boolean>} - A promise that resolves to true if the check passes, false otherwise.
-   */
-  runCheck: () => Promise<boolean>
-  /**
-   * Handles the failure of the startup check.
-   *
-   * @returns {Promise<void> | void} - A promise that resolves when the failure handling is complete.
-   */
-  onFail: () => Promise<void> | void
-  /**
-   * Handles the success of the startup check.
-   *
-   * @returns {Promise<void> | void} - A promise that resolves when the success handling is complete.
-   */
-  onSuccess?: () => Promise<void> | void
-}
-
-// Common utilities used across system checks
-export interface SystemCheckUtils {
-  /**
-   * Dispatch function to send actions to the store.
-   *
-   * @type {Dispatch<ReducerAction<any>>}
-   */
-  dispatch: Dispatch<ReducerAction<any>>
-  /**
-   * Translation function for internationalization.
-   *
-   * @type {TFunction}
-   */
-  translation: TFunction
-  /**
-   * Logger for logging messages and errors.
-   *
-   * @type {BifoldLogger}
-   */
-  logger: BifoldLogger
-}
-
-/**
- * Runs a series of startup checks and handles failures and successes accordingly.
- *
- * @param {SystemCheckStrategy[]} checks - An array of startup checks to run.
- * @returns {*} {Promise<boolean[]>} - An array of boolean results indicating the success of each check.
- */
-export async function runSystemChecks(checks: SystemCheckStrategy[]) {
-  const startupPromises: Promise<boolean>[] = []
-
-  // Add all startup check promises to array
-  for (const check of checks) {
-    startupPromises.push(check.runCheck())
-  }
-
-  // Wait for all startup checks to complete in parallel
-  const results = await Promise.all(startupPromises)
-
-  // Handle failures in order
-  // To be determined if we want automatic failure handling or not (pass param if not)
-  for (let index = 0; index < results.length; index++) {
-    if (!results[index]) {
-      await checks[index].onFail()
-      continue
-    }
-
-    await checks[index].onSuccess?.()
-  }
-
-  return results
-}
-
-/**
- * Checks if the number of registered devices exceeds the allowed limit.
- *
- * Note: Determines the limit based on the values in the ID token.
- *   On failure, it dispatches a warning banner message.
- *   On success, it removes the banner if it exists.
- *
- * @class DeviceCountStartupCheck
- * @implements {SystemCheckStrategy}
- */
-export class DeviceCountSystemCheck implements SystemCheckStrategy {
-  private readonly getIdToken: () => Promise<IdToken>
-  private readonly utils: SystemCheckUtils
-
-  constructor(getIdToken: () => Promise<IdToken>, utils: SystemCheckUtils) {
-    this.getIdToken = getIdToken
-    this.utils = utils
-  }
-
-  /**
-   * Runs the device count check to verify if the number of registered devices is within the allowed limit.
-   *
-   * @returns {Promise<boolean>} - A promise that resolves to true if the device count is within the limit, false otherwise.
-   */
-  async runCheck() {
-    try {
-      const idToken = await this.getIdToken()
-
-      return idToken.bcsc_devices_count < idToken.bcsc_max_devices
-    } catch (error) {
-      this.utils.logger.error('DeviceSystemCheck: Id token request failed', error as Error)
-      return false
-    }
-  }
-
-  /**
-   * Handles the failure of the device count check by dispatching a warning banner message.
-   *
-   * @returns {*} {void}
-   */
-  onFail() {
-    this.utils.dispatch({
-      type: BCDispatchAction.ADD_BANNER_MESSAGE,
-      payload: [
-        {
-          id: BCSCBanner.DEVICE_LIMIT_EXCEEDED,
-          title: this.utils.translation('Unified.SystemChecks.Devices.DeviceLimitReachedBannerTitle'),
-          type: 'warning',
-          variant: 'summary',
-          dismissible: false, // Non-dismissible banner (user must dismiss from screen)
-        },
-      ],
-    })
-  }
-
-  /**
-   * Handles the success of the device count check by removing the warning banner message if it exists.
-   *
-   * @returns {*} {void}
-   */
-  onSuccess() {
-    this.utils.dispatch({ type: BCDispatchAction.REMOVE_BANNER_MESSAGE, payload: [BCSCBanner.DEVICE_LIMIT_EXCEEDED] })
-  }
-}
-
-/**
- * Checks the IAS server status and dispatches banner messages based on availability.
- * Will show error banner if server is unavailable, or info banner if there is a status message.
- *
- * Note: Determines server availability via the config API.
- *   On failure, it dispatches a warning banner message.
- *   On success, it removes the banner if it exists.
- *
- * @class ServerStatusStartupCheck
- * @implements {SystemCheckStrategy}
- */
-export class ServerStatusSystemCheck implements SystemCheckStrategy {
-  private readonly getServerStatus: () => Promise<ServerStatusResponseData>
-  private readonly utils: SystemCheckUtils
-  private serverStatus?: ServerStatusResponseData
-
-  constructor(getServerStatus: () => Promise<ServerStatusResponseData>, utils: SystemCheckUtils) {
-    this.getServerStatus = getServerStatus
-    this.utils = utils
-  }
-
-  /**
-   * Runs the server status check to verify if the IAS server is available.
-   *
-   * @returns {*} {Promise<boolean>} - A promise that resolves to true if the server is available, false otherwise.
-   */
-  async runCheck() {
-    try {
-      const serverStatus = await this.getServerStatus()
-
-      // Store the server status for use in onFail/onSuccess
-      this.serverStatus = serverStatus
-
-      return serverStatus.status === 'ok'
-    } catch (error) {
-      this.utils.logger.error('ServerStatusSystemCheck: Server status request failed', error as Error)
-      return false
-    }
-  }
-
-  /**
-   * Handles the failure of the server status check by dispatching an error banner message.
-   *
-   * @returns {*} {void}
-   */
-  onFail() {
-    this.utils.dispatch({
-      type: BCDispatchAction.ADD_BANNER_MESSAGE,
-      payload: [
-        {
-          id: BCSCBanner.IAS_SERVER_UNAVAILABLE,
-          title:
-            this.serverStatus?.statusMessage ??
-            this.utils.translation('Unified.SystemChecks.ServerStatus.UnavailableBannerTitle'),
-          type: 'error',
-          variant: 'summary',
-          dismissible: true,
-        },
-      ],
-    })
-  }
-
-  /**
-   * Handles the success of the server status check by removing the error banner message if it exists.
-   *
-   * @returns {*} {void}
-   */
-  onSuccess() {
-    this.utils.dispatch({ type: BCDispatchAction.REMOVE_BANNER_MESSAGE, payload: [BCSCBanner.IAS_SERVER_NOTIFICATION] })
-    this.utils.dispatch({ type: BCDispatchAction.REMOVE_BANNER_MESSAGE, payload: [BCSCBanner.IAS_SERVER_UNAVAILABLE] })
-
-    if (!this.serverStatus?.statusMessage) {
-      return
-    }
-
-    // If the server has a status message, show it as an info banner
-    this.utils.dispatch({
-      type: BCDispatchAction.ADD_BANNER_MESSAGE,
-      payload: [
-        {
-          id: BCSCBanner.IAS_SERVER_NOTIFICATION,
-          title: this.serverStatus.statusMessage,
-          type: 'info',
-          variant: 'summary',
-          dismissible: true,
-        },
-      ],
-    })
-  }
-}
